+++ driver/src/main/java/oracle/nosql/driver/Region.java
@@ -59,6 +59,8 @@ public class Region {
     private static final Map<String, Region> OC17_REGIONS = new HashMap<>();
     private static final Map<String, Region> OC19_REGIONS = new HashMap<>();
     private static final Map<String, Region> OC20_REGIONS = new HashMap<>();
+    private static final Map<String, Region> OC22_REGIONS = new HashMap<>();
+    private static final Map<String, Region> OC24_REGIONS = new HashMap<>();
 
     /* OC1 */
     public static final Region AF_JOHANNESBURG_1 = new Region("af-johannesburg-1");
@@ -97,6 +99,7 @@ public class Region {
     public static final Region US_ASHBURN_1 = new Region("us-ashburn-1");
     public static final Region US_PHOENIX_1 = new Region("us-phoenix-1");
     public static final Region US_SANJOSE_1 = new Region("us-sanjose-1");
+    public static final Region US_SALTLAKE_2 = new Region("us-saltlake-2");
     public static final Region US_CHICAGO_1 = new Region("us-chicago-1");
 
     public static final Region CA_MONTREAL_1 = new Region("ca-montreal-1");
@@ -104,6 +107,7 @@ public class Region {
 
     public static final Region SA_SAOPAULO_1 = new Region("sa-saopaulo-1");
     public static final Region SA_SANTIAGO_1 = new Region("sa-santiago-1");
+    public static final Region SA_VALPARAISO_1 = new Region("sa-valparaiso-1");
     public static final Region SA_VINHEDO_1 = new Region("sa-vinhedo-1");
 
     /* OC2 */
@@ -155,6 +159,12 @@ public class Region {
     /* OC20 */
     public static final Region EU_JOVANOVAC_1 = new Region("eu-jovanovac-1");
 
+    /* OC22 */
+    public static final Region EU_DCC_ROME_1 = new Region("eu-dcc-rome-1");
+
+    /* OC24 */
+    public static final Region EU_DCC_ZURICH_1 = new Region("eu-dcc-zurich-1");
+
     static {
         /* OC1 */
         /* AF */
@@ -196,12 +206,14 @@ public class Region {
         /* LAD */
         OC1_REGIONS.put(SA_SAOPAULO_1.getRegionId(), SA_SAOPAULO_1);
         OC1_REGIONS.put(SA_SANTIAGO_1.getRegionId(), SA_SANTIAGO_1);
+        OC1_REGIONS.put(SA_VALPARAISO_1.getRegionId(), SA_VALPARAISO_1);
         OC1_REGIONS.put(SA_VINHEDO_1.getRegionId(), SA_VINHEDO_1);
 
         /* North America */
         OC1_REGIONS.put(US_ASHBURN_1.getRegionId(), US_ASHBURN_1);
         OC1_REGIONS.put(US_PHOENIX_1.getRegionId(), US_PHOENIX_1);
         OC1_REGIONS.put(US_SANJOSE_1.getRegionId(), US_SANJOSE_1);
+        OC1_REGIONS.put(US_SALTLAKE_2.getRegionId(), US_SALTLAKE_2);
         OC1_REGIONS.put(US_CHICAGO_1.getRegionId(), US_CHICAGO_1);
 
         OC1_REGIONS.put(CA_MONTREAL_1.getRegionId(), CA_MONTREAL_1);
@@ -255,6 +267,12 @@ public class Region {
 
         /* OC20 */
         OC20_REGIONS.put(EU_JOVANOVAC_1.getRegionId(), EU_JOVANOVAC_1);
+
+        /* OC22 */
+        OC22_REGIONS.put(EU_DCC_ROME_1.getRegionId(), EU_DCC_ROME_1);
+
+        /* OC24 */
+        OC24_REGIONS.put(EU_DCC_ZURICH_1.getRegionId(), EU_DCC_ZURICH_1);
     }
 
     private final static MessageFormat OC1_EP_BASE = new MessageFormat(
@@ -281,6 +299,10 @@ public class Region {
         "https://nosql.{0}.oci.oraclecloud.eu");
     private final static MessageFormat OC20_EP_BASE = new MessageFormat(
         "https://nosql.{0}.oci.oraclecloud20.com");
+    private final static MessageFormat OC22_EP_BASE = new MessageFormat(
+        "https://nosql.{0}.oci.oraclecloud22.com");
+    private final static MessageFormat OC24_EP_BASE = new MessageFormat(
+        "https://nosql.{0}.oci.oraclecloud24.com");
 
     private String regionId;
 
@@ -328,6 +350,12 @@ public class Region {
         }
         if (isOC20Region(regionId)) {
             return OC20_EP_BASE.format(new Object[] { regionId });
+        }
+        if (isOC22Region(regionId)) {
+            return OC22_EP_BASE.format(new Object[] { regionId });
+        }
+        if (isOC24Region(regionId)) {
+            return OC24_EP_BASE.format(new Object[] { regionId });
         }
         throw new IllegalArgumentException(
             "Unable to find endpoint for unknwon region" + regionId);
@@ -379,6 +407,12 @@ public class Region {
         }
         if (region == null) {
             region = OC20_REGIONS.get(regionId);
+        }
+        if (region == null) {
+            region = OC22_REGIONS.get(regionId);
+        }
+        if (region == null) {
+            region = OC24_REGIONS.get(regionId);
         }
 
         return region;
@@ -511,6 +545,26 @@ public class Region {
      */
     public static boolean isOC20Region(String regionId) {
         return (OC20_REGIONS.get(regionId) != null);
+    }
+
+    /**
+     * @hidden
+     * Internal use only
+     * @param regionId the region id
+     * @return the value
+     */
+    public static boolean isOC22Region(String regionId) {
+        return (OC22_REGIONS.get(regionId) != null);
+    }
+
+    /**
+     * @hidden
+     * Internal use only
+     * @param regionId the region id
+     * @return the value
+     */
+    public static boolean isOC24Region(String regionId) {
+        return (OC24_REGIONS.get(regionId) != null);
     }
 
     /**